+++ src/test/java/com/baro/auth/fake/jwt/FakeTokenStorage.java
@@ -30,13 +30,13 @@ public class FakeTokenStorage implements TokenStorage {
 
     @Override
     public String findRefreshToken(String key) {
-        if (!tokens.containsKey("RT:" + key))
-            throw new AuthException(AuthExceptionType.REFRESH_TOKEN_DOES_NOT_EXIST);
-
-        if (tokens.get("RT:" + key).expireTime().isAfter(timeServer.now()))
+        if (tokens.get("RT:" + key).expireTime().isAfter(timeServer.now())){
             return tokens.get("RT:" + key).token();
-
-        else throw new JwtTokenException(JwtTokenExceptionType.EXPIRED_JWT_TOKEN);
+        }
+        if (!tokens.containsKey("RT:" + key)) {
+            throw new AuthException(AuthExceptionType.REFRESH_TOKEN_DOES_NOT_EXIST);
+        }
+        throw new JwtTokenException(JwtTokenExceptionType.EXPIRED_JWT_TOKEN);
     }
 
     record TokenInfo(String token, Instant expireTime) {
