+++ src/test/java/com/baro/auth/presentation/oauth/OAuthApiDocumentTest.java
@@ -13,8 +13,6 @@ import static org.springframework.restdocs.request.RequestDocumentation.pathPara
 import static org.springframework.restdocs.request.RequestDocumentation.queryParameters;
 
 import com.baro.auth.application.AuthService;
-import com.baro.auth.application.dto.SignInDto;
-import com.baro.auth.domain.Token;
 import com.baro.auth.infra.oauth.kakao.KakaoOAuthClient;
 import com.baro.auth.infra.oauth.kakao.KakaoRequestApi;
 import com.baro.auth.infra.oauth.kakao.dto.KakaoMemberResponse;
@@ -90,38 +88,6 @@ class OAuthApiDocumentTest extends RestApiDocumentationTest {
                 ))
                 .queryParam("authCode", "authCode")
                 .when().get(url, "kakao")
-                .then().log().all()
-                .extract();
-
-        // then
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
-    }
-
-    @Test
-    void reissue() {
-        // given
-        var url = "/auth/reissue";
-        when(kakaoRequestApi.requestToken(anyMap()))
-                .thenReturn(new KakaoTokenResponse("Bearer", "accessToken", "idToken",
-                        1000, "refreshToken", 1000, "scope"));
-        when(kakaoRequestApi.requestMemberInfo(anyString()))
-                .thenReturn(new KakaoMemberResponse(1L, "nickname",
-                        new KakaoMemberResponse.Properties("nickname"),
-                        new KakaoAccount(false, new KakaoAccount.Profile("nickname"), "email")
-                ));
-        SignInDto dto = new SignInDto("name", "email", "1", "kakao");
-        Token token = authService.signIn(dto);
-
-        // when
-        var response = given(requestSpec).log().all()
-                .filter(document(DEFAULT_REST_DOCS_PATH,
-                        queryParameters(
-                                parameterWithName("refreshToken").description("refreshToken")
-                        )
-                ))
-                .header(HttpHeaders.AUTHORIZATION, "Bearer " + token.accessToken())
-                .queryParam("refreshToken", token.refreshToken())
-                .when().get(url)
                 .then().log().all()
                 .extract();
